streamlit-llm-otel/requirements.txt
streamlit
open-telemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
requests
pandas
numpy